moved property panel to correct folder, added secondary property panel
