+++ Assets/UI Toolkit/Resources/UI/AppRoot.uxml
@@ -21,10 +21,8 @@
                 <ui:VisualElement picking-mode="Ignore" style="flex-grow: 1;"/>
             </ui:VisualElement>
             <ui:VisualElement name="PropertiesSection" class="properties-section">
-                <nl3d:ContentContainer leading-icon="Paint" text="Kleur aanpassen" show-close-button="true">
-                    <nl3d:ColorPicker name="PropertiesColorPicker" class="properties-section-column"/>
-                </nl3d:ContentContainer>
-                <nl3d:PropertiesPanel name="PropertiesPanel" class="properties-section-column"/>
+                <nl3dP:SecondaryPropertiesPanel />
+                <nl3dP:PropertiesPanel name="PropertiesPanel" class="properties-section-column"/>
             </ui:VisualElement>
         </ui:VisualElement>
         <ui:VisualElement name="BottomSection" picking-mode="Ignore" style="align-items: flex-end; flex-direction: row;">
