+++ Assets/UI Toolkit/Resources/UI/Panels/SecondaryPropertiesPanel-style.uss
@@ -1,0 +1,57 @@
+.properties-panel {
+    --properties-bg-color: var(--color-white);
+    --properties-border-radius: var(--spacing-2);
+    --properties-width: 400px;
+    justify-content: space-between;
+    flex-shrink: 0;
+    flex-grow: 0;
+    align-self: stretch;
+    min-width: 320px;
+    width: var(--properties-width);
+    max-width: 520px;
+    margin-left: 0px;
+    margin-top: 0px;
+    margin-right: 0px;
+    margin-bottom: var(--spacing-20);
+    padding: var(--spacing-5);
+    padding-top: var(--spacing-4);
+    padding-right: var(--spacing-5);
+    padding-bottom: var(--spacing-5);
+    background-color: var(--properties-bg-color);
+    border-radius: var(--properties-border-radius);
+}
+
+.properties-panel .properties-panel__header {
+    flex-direction: row;
+    flex-shrink: 0;
+    align-content: stretch;
+    align-items: center;
+    margin-bottom: var(--spacing-4);
+    padding: 0;
+}
+
+.properties-panel .properties-panel__header-title {
+    flex-grow: 1;
+    flex-shrink: 0;
+}
+
+.properties-panel .properties-panel__header-button-close {
+    flex-grow: 0;
+    flex-shrink: 0;
+    width: var(--spacing-6);
+    height: var(--spacing-6);
+    align-items: center;
+    flex-wrap: nowrap;
+    align-content: center;
+    border-width: 0;
+    background-size: cover;
+}
+.properties-panel .properties-panel__header-button-close .icon {
+    background-size: var(--spacing-6);
+}
+.properties-panel .properties-panel__content {
+    
+    flex-grow: 1;
+    flex-shrink: 1;
+    padding: 0px;
+}
