+++ Assets/UI Toolkit/Resources/UI/Panels/SecondaryPropertiesPanel-style.uss.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d3c508949644340deb14138861572e15
+ScriptedImporter:
+  internalIDToNameTable: []
+  externalObjects: {}
+  serializedVersion: 2
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+  script: {fileID: 12385, guid: 0000000000000000e000000000000000, type: 0}
+  disableValidation: 0
