+++ Assets/UI Toolkit/Resources/UI/Panels/SecondaryPropertiesPanel.uxml.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 6cb7594b3a37247d686d2efa3bd98f52
+ScriptedImporter:
+  internalIDToNameTable: []
+  externalObjects: {}
+  serializedVersion: 2
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+  script: {fileID: 13804, guid: 0000000000000000e000000000000000, type: 0}
