+++ Assets/UI Toolkit/Scripts/Panels/PropertiesPanel.cs
@@ -1,8 +1,10 @@
 using Netherlands3D.UI_Toolkit;
+using Netherlands3D.UI.Components;
 using Netherlands3D.UI.ExtensionMethods;
 using UnityEngine.UIElements;
+using Button = UnityEngine.UIElements.Button;
 
-namespace Netherlands3D.UI.Components
+namespace Netherlands3D.UI.Panels
 {
     [UxmlElement]
     public partial class PropertiesPanel : VisualElement
