+++ Assets/UI Toolkit/Scripts/Panels/SecondaryPropertiesPanel.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c6a6c5b4ab4d14f8f8f66a554cc59b95
